requirements.txt
pandas
requests
python-dotenv      # for reading .env (API key)
pyarrow            # fast parquet/feather IO
